refactor: criando arquivos index para exportar recursos do core
