+++ projects/backend/src/core/session/@shared/session-token/index.ts
@@ -1,1 +1,2 @@
 export type { SessionToken, SignOutput, VerifyOutput } from './session-token'
+export { JwtSessionToken } from './jwt.session-token'
