+++ projects/backend/src/core/session/index.ts
@@ -1,0 +1,4 @@
+export { Session } from './entity/session.entity'
+export * from './repository'
+export * from './use-cases'
+export { type SessionToken, JwtSessionToken } from './@shared/session-token'
