+++ projects/backend/src/core/session/use-cases/refresh/index.ts
@@ -1,0 +1,3 @@
+export { RefreshUseCase } from './refresh.use-case'
+export { RefreshInput } from './refresh.input'
+export { RefreshOutput } from './refresh.output'
