+++ projects/backend/src/core/session/use-cases/sign-in/index.ts
@@ -1,0 +1,3 @@
+export { SignInUseCase } from './sign-in.use-case'
+export { SignInInput } from './sign-in.input'
+export { SignInOutput } from './sign-in.output'
