+++ projects/backend/src/core/session/use-cases/sign-out/index.ts
@@ -1,0 +1,3 @@
+export { SignOutUseCase } from './sign-out.use-case'
+export { SignOutInput } from './sign-out.input'
+export { SignOutOutput } from './sign-out.output'
