+++ projects/backend/src/core/session/use-cases/sign-out/sign-out.output.ts
