+++ projects/backend/src/core/user/index.ts
@@ -1,0 +1,2 @@
+export { User } from './entity/user.entity'
+export * from './repository'
