+++ projects/backend/src/core/user/repository/index.ts
@@ -1,0 +1,3 @@
+export type { UserRepository } from './user.repository'
+export { UserMysqlRepository } from './user.mysql.repository'
+export { UserInMemoryRepository } from './user.in-memory.repository'
